Extract workspace options into workspace/options.py
