+++ tsrc/cli/init.py
@@ -12,7 +12,7 @@ def main(args: argparse.Namespace) -> None:
     workspace_path = args.workspace_path or os.getcwd()
     workspace = tsrc.workspace.Workspace(Path(workspace_path))
     ui.info_1("Configuring workspace in", ui.bold, workspace_path)
-    manifest_options = tsrc.workspace.options_from_args(args)
+    manifest_options = tsrc.workspace.options.options_from_args(args)
     workspace.configure_manifest(manifest_options)
     workspace.load_manifest()
     workspace.clone_missing()
