+++ tsrc/workspace/__init__.py
@@ -1,18 +1,12 @@
-""" Tools to manage tsrc workspaces
-
-Mostly used by tsrc/cli.py
-
+""" Implementation of the tsrc Workspace: a collection of git repositories
 """
 
-import argparse
 import stat
 import textwrap
 from typing import cast, Iterable, List, Tuple, Dict, Any, Optional, NewType  # noqa
 
-import attr
 from path import Path
 import ruamel.yaml
-import schema
 import ui
 
 import tsrc
@@ -20,43 +14,8 @@ import tsrc.executor
 import tsrc.git
 import tsrc.manifest
 
-
-OPTIONS_SCHEMA = schema.Schema({
-    "url": str,
-    schema.Optional("branch"): str,
-    schema.Optional("tag"): str,
-    schema.Optional("groups"): [str],
-    schema.Optional("shallow"): bool,
-})
-
-
-@attr.s
-class Options:
-    url = attr.ib(default=None)  # type: str
-    branch = attr.ib(default="master")  # type: str
-    tag = attr.ib(default=None)  # type: Optional[str]
-    shallow = attr.ib(default=False)  # type: bool
-    groups = attr.ib(default=list())  # type: List[str]
-
-
-def options_from_dict(as_dict: dict) -> Options:
-    res = Options()
-    res.url = as_dict["url"]
-    res.branch = as_dict.get("branch", "master")
-    res.tag = as_dict.get("tag")
-    res.shallow = as_dict.get("shallow", False)
-    res.groups = as_dict.get("groups") or list()
-    return res
-
-
-def options_from_args(args: argparse.Namespace) -> Options:
-    as_dict = vars(args)  # type: dict
-    return options_from_dict(as_dict)
-
-
-def options_from_file(cfg_path: Path) -> Options:
-    as_dict = tsrc.config.parse_config_file(cfg_path, OPTIONS_SCHEMA)  # type: dict
-    return options_from_dict(as_dict)
+from .options import Options
+from .options import options_from_file
 
 
 class LocalManifest:
@@ -208,11 +167,6 @@ class Workspace():
         return self.local_manifest.shallow
 
     def clone_missing(self) -> None:
-        """ Clone missing repos.
-
-        Called at the beginning of `tsrc init` and `tsrc sync`
-
-        """
         to_clone = list()
         for repo in self.get_repos():
             repo_path = self.joinpath(repo.src)
